% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranslaSeq-package.R
\docType{package}
\name{TranslaSeq-package}
\alias{TranslaSeq-package}
\title{TranslaSeq: A package for the assessment of translational control.}
\description{
The TranslaSeq package provides the TranslaSeq function.
}
\seealso{
\code{\link{TranslaSeq}}\cr\cr
         A working example can be found at
         \url{https://franciscodavid.github.io/TranslaSeq/vignette.html}
}
\author{
Francisco D. Morón-Duran
}
