% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranslaSeq.R
\name{TranslaSeq}
\alias{TranslaSeq}
\title{Translational control assessment from ribosome footprint and total RNA
libraries}
\usage{
TranslaSeq(metadata, refname, fafile, gtffile, ctrlabel, condition,
           outdir = 'TranslaSeq.out', preprocess = F, threads = 1,
           Radapt = 'CTGTAGGCACCATCAAT', platform = 'Illumina', verbose = F)
}
\arguments{
\item{metadata}{A dataframe with samples metadata (see Details section).}

\item{refname}{Name given to the genome annotation used in the analysis.}

\item{fafile}{Filepath or URL address of the genome sequence FASTA file.}

\item{gtffile}{Filepath or URL address of the genome annotation GTF file.}

\item{ctrlabel}{Text label for the control condition in metadata dataframe.}

\item{condition}{Column name containing the condition variable in metadata.}

\item{outdir}{Path in which output results will be stored.}

\item{preprocess}{Boolean indicating whether FASTQ input should be
preprocessed to remove adaptors.}

\item{threads}{Number of threads to be used in processing the input data.}

\item{Radapt}{Adaptor sequence to be removed from the input FASTQ files.}

\item{platform}{Name of the platform which generated input data (default:
Illumina).}

\item{verbose}{A boolean used for debugging the entire pipeline.}
}
\value{
A dataframe specifying mean ribosome footprint counts, total RNA
        counts and translation efficiency ratio for the control samples along
        with log2 fold changes for the case samples and their translation
        efficiency p-value and adjusted p-value.
}
\description{
This function compares gene counts from ribosome footprint libraries with
gene counts from total RNA libraries for the same samples across different
experimental conditions to assess translational efficiency changes for
individual transcripts and their statistical significance.
}
\details{
In a metadata dataframe each row represents an input file. This data
         structure has the following mandatory columns: \cr
         \enumerate{
           \item name: the sample name.
           \item file: the path to the input file.
           \item type: the library type, either 'rna' or 'rpf'.
           \item comment: string describing the input file.
         }
         All other wanted data fields must be inserted between type and
         comment columns. To sum up, the first column in metadata dataframe
         must be 'name'; the second one 'file'; the third one 'type'; then
         any number of arbitrary columns with other data fields and the last
         column must be 'comment'.

         Valid file names should contain these suffixes: \cr
         \itemize{
           \item .fastq[.gz] files providing [compressed] FASTQ data.
           \item .bam|.sam files providing alignment data.
           \item .tsv|.count files providing count values.
         }
         All files from a metadata dataframe must be of the same type.

         From the arbitrary columns, at least one should be named as the
         'condition' argument. In its field values, at least one sample
         must has the label 'ctrlabel', which will be the control condition.
         All other labels different from 'ctrlabel' will be treated as case
         conditions to be compared against the control one.
}
\seealso{
A working example can be found at
         \url{https://franciscodavid.github.io/TranslaSeq/vignette.html}
}
\author{
Francisco D. Morón-Duran
}
